[ALPS07843652] adsp: extend feature-control to 64 bit
For more feature control support,
extend control bit from 32bits to 64bits

MTK-Commit-Id: be6ae57dd37f30e93f6922c7ad825068a364c893

CR-Id: ALPS07843652
Feature: ADSP AP Driver

+++ drivers/misc/mediatek/adsp/adsp_feature_define.h
@@ -12,7 +12,7 @@ enum {
 
 struct adsp_feature_control {
 	int total;
-	unsigned int feature_set; /* from dts */
+	u64 feature_set; /* from dts */
 	struct mutex lock;
 	struct workqueue_struct *wq;
 	struct delayed_work suspend_work;
@@ -30,7 +30,7 @@ bool flush_suspend_work(u32 cid);
 int _adsp_register_feature(u32 cid, u32 fid, u32 opt);
 int _adsp_deregister_feature(u32 cid, u32 fid, u32 opt);
 
-int init_adsp_feature_control(u32 cid, u32 feature_set, int delay_ms,
+int init_adsp_feature_control(u32 cid, u64 feature_set, int delay_ms,
 			struct workqueue_struct *wq,
 			int (*_suspend)(void),
 			int (*_resume)(void));
